[INFRA] split argumentparsing into header and object files
