+++ include/argument_parsing/top_level.hpp
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <argument_parsing/shared.hpp>
+
+void init_top_level_parser(seqan3::argument_parser & parser);
