+++ src/CMakeLists.txt
@@ -6,26 +6,46 @@ cmake_minimum_required (VERSION 3.8)
 #                                                      |
 # raptor_minimiser_model_lib +-----> raptor_search_lib +
 
+# shared interface
+add_library ("${PROJECT_NAME}_interface" INTERFACE)
+target_link_libraries ("${PROJECT_NAME}_interface" INTERFACE seqan3::seqan3)
+target_include_directories ("${PROJECT_NAME}_interface" INTERFACE ../include)
+target_include_directories ("${PROJECT_NAME}_interface" INTERFACE ../lib/robin-hood-hashing/src/include)
+
 # raptor build
 add_library ("${PROJECT_NAME}_build_lib" STATIC raptor_build.cpp)
-target_link_libraries ("${PROJECT_NAME}_build_lib" PUBLIC seqan3::seqan3)
-target_include_directories ("${PROJECT_NAME}_build_lib" PUBLIC ../include)
-target_include_directories ("${PROJECT_NAME}_build_lib" PUBLIC ../lib/robin-hood-hashing/src/include)
+target_link_libraries ("${PROJECT_NAME}_build_lib" PUBLIC "${PROJECT_NAME}_interface")
 
 # the minimiser model will be used for raptor search
 add_library ("${PROJECT_NAME}_minimiser_model_lib" STATIC minimiser_model.cpp)
-target_link_libraries ("${PROJECT_NAME}_minimiser_model_lib" PUBLIC seqan3::seqan3)
-target_include_directories ("${PROJECT_NAME}_minimiser_model_lib" PUBLIC ../include)
+target_link_libraries ("${PROJECT_NAME}_minimiser_model_lib" PUBLIC "${PROJECT_NAME}_interface")
 
 # raptor search
 add_library ("${PROJECT_NAME}_search_lib" STATIC raptor_search.cpp)
 target_link_libraries ("${PROJECT_NAME}_search_lib" PUBLIC "${PROJECT_NAME}_minimiser_model_lib")
 
+# argument parsing
+add_library ("${PROJECT_NAME}_argument_parsing_shared_lib" STATIC argument_parsing/shared.cpp)
+target_link_libraries ("${PROJECT_NAME}_argument_parsing_shared_lib" PUBLIC "${PROJECT_NAME}_interface")
+
+add_library ("${PROJECT_NAME}_argument_parsing_build_lib" STATIC argument_parsing/build.cpp)
+target_link_libraries ("${PROJECT_NAME}_argument_parsing_build_lib" PUBLIC "${PROJECT_NAME}_argument_parsing_shared_lib")
+
+add_library ("${PROJECT_NAME}_argument_parsing_search_lib" STATIC argument_parsing/search.cpp)
+target_link_libraries ("${PROJECT_NAME}_argument_parsing_search_lib" PUBLIC "${PROJECT_NAME}_argument_parsing_shared_lib")
+
+add_library ("${PROJECT_NAME}_argument_parsing_top_level_lib" STATIC argument_parsing/top_level.cpp)
+target_link_libraries ("${PROJECT_NAME}_argument_parsing_top_level_lib" PUBLIC "${PROJECT_NAME}_argument_parsing_shared_lib")
+
 # raptor consists of raptor build and raptor search
-add_library ("${PROJECT_NAME}_lib" STATIC raptor.cpp)
-target_link_libraries ("${PROJECT_NAME}_lib" PUBLIC "${PROJECT_NAME}_build_lib")
-target_link_libraries ("${PROJECT_NAME}_lib" PUBLIC "${PROJECT_NAME}_search_lib")
+add_library ("${PROJECT_NAME}_lib" INTERFACE)
+target_link_libraries ("${PROJECT_NAME}_lib" INTERFACE "${PROJECT_NAME}_argument_parsing_build_lib")
+target_link_libraries ("${PROJECT_NAME}_lib" INTERFACE "${PROJECT_NAME}_argument_parsing_search_lib")
+target_link_libraries ("${PROJECT_NAME}_lib" INTERFACE "${PROJECT_NAME}_argument_parsing_shared_lib")
+target_link_libraries ("${PROJECT_NAME}_lib" INTERFACE "${PROJECT_NAME}_argument_parsing_top_level_lib")
+target_link_libraries ("${PROJECT_NAME}_lib" INTERFACE "${PROJECT_NAME}_build_lib")
+target_link_libraries ("${PROJECT_NAME}_lib" INTERFACE "${PROJECT_NAME}_search_lib")
 
 # the executable
 add_executable ("${PROJECT_NAME}" raptor_main.cpp)
-target_link_libraries ("${PROJECT_NAME}" "${PROJECT_NAME}_lib")
+target_link_libraries ("${PROJECT_NAME}" PUBLIC "${PROJECT_NAME}_lib")
