+++ src/argument_parsing/shared.cpp
@@ -1,0 +1,58 @@
+#include <argument_parsing/shared.hpp>
+
+void init_shared_meta(seqan3::argument_parser & parser)
+{
+    parser.info.app_name = "Raptor";
+    parser.info.author = "Enrico Seiler";
+    parser.info.citation = "Seiler, E. et al. (2020). Raptor: A fast and space-efficient pre-filter for"
+                           " querying very large collections of nucleotide sequences. bioRxiv 2020.10.08.330985. doi:"
+                           " https://doi.org/10.1101/2020.10.08.330985";
+    parser.info.date = "27-05-2021";
+    parser.info.email = "enrico.seiler@fu-berlin.de";
+    parser.info.long_copyright = R"(BSD 3-Clause License
+
+Copyright (c) 2021, Enrico Seiler
+All rights reserved.
+
+Redistribution and use in source and binary forms, with or without
+modification, are permitted provided that the following conditions are met:
+
+1. Redistributions of source code must retain the above copyright notice, this
+   list of conditions and the following disclaimer.
+
+2. Redistributions in binary form must reproduce the above copyright notice,
+   this list of conditions and the following disclaimer in the documentation
+   and/or other materials provided with the distribution.
+
+3. Neither the name of the copyright holder nor the names of its
+   contributors may be used to endorse or promote products derived from
+   this software without specific prior written permission.
+
+THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS \"AS IS\"
+AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
+FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
+OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.)";
+    parser.info.short_copyright = "BSD 3-Clause License";
+    parser.info.short_description = "A fast and space-efficient pre-filter for querying very large collections of nucleotide sequences.";
+    parser.info.url = "https://github.com/seqan/raptor";
+    parser.info.version = "1.2.0";
+}
+
+void try_parsing(seqan3::argument_parser & parser)
+{
+    try
+    {
+        parser.parse();
+    }
+    catch (seqan3::argument_parser_error const & ext)
+    {
+        std::cerr << "[Error] " << ext.what() << '\n';
+        std::exit(-1);
+    }
+}
